+++ functions/functions-elements.php
@@ -1,13 +1,25 @@
-<?php
+<?
 /**
  * Functions to output ACF flexible elements.
  * WPSeed_Text and WPSeed_Gallery are placeholder presets.
  *
  * @author     Flurin Dürst
- * @version    1.1
+ * @version    1.2
  * @since      WPSeed 0.10.0
  *
  */
+
+/* ADD TITLES TO ELEMENT BARS
+/----------------------------*/
+// adds the title-sub_field to the ACF-element bar. Edit `name` at `add_filter` to match your ACF-value.
+// » https://www.advancedcustomfields.com/resources/acf-fields-flexible_content-layout_title/
+function acf_flexiblecontent_title( $title, $field, $layout, $i ) {
+  if (!empty(get_sub_field('title'))) {          // remove layout title from text
+  	$title = get_sub_field('title')." ($title)";  // add new title
+  }
+  return $title;
+}
+add_filter('acf/fields/flexible_content/layout_title/name=element', 'acf_flexiblecontent_title', 10, 4);
 
 
 /* GATHER ELEMENTS
@@ -29,10 +41,10 @@ function elements() {
 function WPSeed_Text() {
   ob_start('sanitize_output') ?>
   <section class="text">
-    <h2><?php the_sub_field('title') ?></h2>
-    <?php the_sub_field('content') ?>
+    <h2><? the_sub_field('title') ?></h2>
+    <? the_sub_field('content') ?>
   </section>
-  <?php return ob_get_flush();
+  <? return ob_get_flush();
 }
 
 /* GALLERY
@@ -40,16 +52,16 @@ function WPSeed_Text() {
 function WPSeed_Gallery() {
   ob_start('sanitize_output') ?>
     <section class="gallery">
-      <h2><?php the_sub_field('title') ?></h2>
-      <?php if (have_rows('galleryimg')): ?>
+      <h2><? the_sub_field('title') ?></h2>
+      <? if (have_rows('galleryimg')): ?>
         <ul>
-          <?php while (have_rows('galleryimg')) : the_row(); ?>
-            <?php $img = get_sub_field('img') ?>
+          <? while (have_rows('galleryimg')) : the_row(); ?>
+            <? $img = get_sub_field('img') ?>
             <li><img src="<?= $img['url'] ?>" /></li>
-          <?php endwhile ?>
+          <? endwhile ?>
         </ul>
-      <?php endif ?>
+      <? endif ?>
     </section>
-  <?php return ob_get_flush();
+  <? return ob_get_flush();
 }
 ?>